third_party/blink/web_tests/wpt_internal/fenced_frame/disallowed-navigations-dangling-markup-urn.https.html

<!DOCTYPE html>
<title>Fenced frame disallowed navigations with potentially-dangling markup</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>
<script src="/fetch/private-network-access/resources/support.sub.js"></script>
<script src="resources/dangling-markup-helper.js"></script>

<body>

<script>
// These tests assert that fenced frames cannot be navigated to a urn:uuid URL
// that represents an HTTPS URLs with dangling markup.
for (const substring of kDanglingMarkupSubstrings) {
  promise_test(async t => {
    const key = token();

    // Copied from from `generateURN()`, since we have to modify the final URL
    // that goes into `selectURL`.
    try {
      await sharedStorage.worklet.addModule(
        "/wpt_internal/shared_storage/resources/simple-module.js");
    } catch (e) {
      // See documentation in `generateURN()`.
    }

    let url_string = generateURL("resources/report-url.html?blocked", [key]).toString();
    url_string = url_string.replace("blocked", substring);

    const urn = await sharedStorage.selectURL(
      "test-url-selection-operation", [{url: url_string}], {data: {'mockResult': 0}}
    );

    const fencedframe = attachFencedFrame(urn, /*mode=*/'opaque-ads');
    const loaded_promise = nextValueFromServer(key);
    const result = await Promise.any([loaded_promise, getTimeoutPromise(t)]);
    assert_equals(result, "NOT LOADED");
  }, `fenced frame opaque URN => https: URL with dangling markup '${substring}'`);
}

</script>

</body>
